+++ packages/perseus/src/widgets/interactive-graphs/interactive-graph-question-builder.test.ts
@@ -198,7 +198,7 @@ describe("InteractiveGraphQuestionBuilder", () => {
                 graph: {
                     type: "segment",
                     numSegments: 1,
-                    coords: [
+                    startCoords: [
                         [
                             [0, 0],
                             [2, 2],
@@ -257,7 +257,7 @@ describe("InteractiveGraphQuestionBuilder", () => {
             expect.objectContaining({
                 graph: {
                     type: "linear",
-                    coords: [
+                    startCoords: [
                         [3, 0],
                         [3, 3],
                     ],
@@ -328,7 +328,7 @@ describe("InteractiveGraphQuestionBuilder", () => {
             expect.objectContaining({
                 graph: {
                     type: "linear-system",
-                    coords: [
+                    startCoords: [
                         [
                             [-3, 0],
                             [-3, 3],
@@ -393,7 +393,7 @@ describe("InteractiveGraphQuestionBuilder", () => {
             expect.objectContaining({
                 graph: {
                     type: "ray",
-                    coords: [
+                    startCoords: [
                         [3, 0],
                         [3, 3],
                     ],
@@ -429,7 +429,7 @@ describe("InteractiveGraphQuestionBuilder", () => {
         const graph = question.widgets["interactive-graph 1"];
         expect(graph.options).toEqual(
             expect.objectContaining({
-                graph: {type: "circle", center: [9, 9], radius: 5},
+                graph: {type: "circle", startCoords: [9, 9], radius: 5},
                 correct: {type: "circle", radius: 3, center: [-1, -1]},
             }),
         );
@@ -475,7 +475,7 @@ describe("InteractiveGraphQuestionBuilder", () => {
             expect.objectContaining({
                 graph: {
                     type: "quadratic",
-                    coords: [
+                    startCoords: [
                         [-1, -1],
                         [0, 0],
                         [1, -1],
@@ -530,7 +530,7 @@ describe("InteractiveGraphQuestionBuilder", () => {
             expect.objectContaining({
                 graph: {
                     type: "sinusoid",
-                    coords: [
+                    startCoords: [
                         [0, 0],
                         [1, -1],
                     ],
@@ -652,7 +652,7 @@ describe("InteractiveGraphQuestionBuilder", () => {
                 graph: {
                     type: "point",
                     numPoints: 2,
-                    coords: [
+                    startCoords: [
                         [3, 3],
                         [4, 4],
                     ],
@@ -715,7 +715,7 @@ describe("InteractiveGraphQuestionBuilder", () => {
             expect.objectContaining({
                 graph: {
                     type: "angle",
-                    coords: [
+                    startCoords: [
                         [3, 3],
                         [4, 4],
                         [5, 5],