+++ packages/perseus/src/widgets/interactive-graphs/stateful-mafs-graph.tsx
@@ -99,6 +99,7 @@ export const StatefulMafsGraph = React.forwardRef<
     const snapTo = graph.type === "polygon" ? graph.snapTo : null;
     const showAngles = graph.type === "polygon" ? graph.showAngles : null;
     const showSides = graph.type === "polygon" ? graph.showSides : null;
+    const startCoords = graph.startCoords ?? null;
 
     const originalPropsRef = useRef(props);
     const latestPropsRef = useLatestRef(props);
@@ -117,7 +118,9 @@ export const StatefulMafsGraph = React.forwardRef<
         snapTo,
         showAngles,
         showSides,
+        startCoords,
         latestPropsRef,
+        startCoords,
     ]);
 
     return <MafsGraph {...props} state={state} dispatch={dispatch} />;